renamed openmdao.math to openmdao.jax to make it obvious that it is dependent upon jax. reorganized some of the functions within openmdao.jax. updated documentation.
