+++ openmdao/jax/__init__.py
@@ -5,4 +5,5 @@ These functions are implemented using jax.numpy for the underlying calculations.
 This will allow these functions to be used in a way that supports jax's automatic differentiation
 and just-in-time compilation capabilities.
 """
-from .jax_functions import act_tanh, ks_max, ks_min, smooth_max, smooth_min, smooth_abs
+from .smooth import act_tanh, smooth_max, smooth_min, smooth_abs
+from .ks import ks_max, ks_min
